requirements.txt
ccxt
python-dotenv